requirements.txt
Appium_Python_Client==2.11.1
clean-text==1.1.4
selenium==4.11.2
